pyproject.toml
[project]
name = "yambler"
version = "0.1.0"
description = "build yaml from a manifest"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyaml>=25.7.0",
    "pyyaml>=6.0.3",
    "tomli>=2.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]

[tool.setuptools.packages.find]
include = ["yambler"]
